Rename food/fatsecret/OAuth.py -> food/fatsecret/OAuthModels.py
When running platemate in Docker on MacOS with the repo code
mounted as a voulume, python fails to import fatsectet modules
because the mounted filesystem is case-preserving but case-insencitive (go figure)
and the import statements
```
from oauth import oauth
from OAuth import *
```
are both treated as importing the same module `oauth`.

Renaming OAuth.py to OAuthModels.py solves this.
